+++ features/Templates/tests/templates-smoke.php
@@ -64,7 +64,6 @@ $appData = [ './nino/uid' => $sandbox ];
 $appData['./nino/filesystem/path'] = $sandbox;
 $appData['./nino/filesystem/configpath'] = $sandbox. '/private';
 $appData['./nino/filesystem/contentpath'] = $sandbox. '/private';
-$appData['./nino/filesystem/privatepath'] = $sandbox. '/private';
 $appData['./nino/filesystem/publicpath'] = $sandbox. '/public';
 $appData['/nino/dir'] = '';
 $appData['/nino/locales/native'] = 'en_US';